Added Field Definitions function
Added button on data.php and separate functions on data.js.

Created a separate sql file to create table and store on database. Used similar methods done in the other two tables to pull data from database.

So far, the table is not being populated but it worked before so I feel like the issue is minor. Will take a closer look tomorrow.